+++ packages/language-server/CHANGELOG.md
@@ -1,3 +1,59 @@
+## 1.2.8 (2024-07-28)
+
+
+### 🩹 Fixes
+
+- update language server to fix #182
+
+- update vscode-css-languageservice
+
+
+### 🧱 Updated Dependencies
+
+- Updated @somesass/language-services to 1.0.7
+
+## 1.2.7 (2024-07-25)
+
+
+### 🩹 Fixes
+
+- update language server for completion regression fixes
+
+- don't crash on ENOENT when looking up realPath
+
+
+### 🧱 Updated Dependencies
+
+- Updated @somesass/language-services to 1.0.6
+
+## 1.2.6 (2024-06-16)
+
+
+### 🩹 Fixes
+
+- update language server for completions bugfix
+
+
+### 🧱 Updated Dependencies
+
+- Updated @somesass/language-services to 1.0.5
+
+## 1.2.5 (2024-06-15)
+
+
+### 🩹 Fixes
+
+- read setting for maximum color decorators ([25d276b5](https://github.com/wkillerud/some-sass/commit/25d276b5))
+
+
+### 🧱 Updated Dependencies
+
+- Updated @somesass/language-services to 1.0.4
+
+## 1.2.4 (2024-05-30)
+
+This was a version bump only for some-sass-language-server to align it with other projects, there were no code changes.
+
 ## 1.2.3 (2024-05-01)