+++ packages/vscode-css-languageservice/CHANGELOG.md
@@ -1,3 +1,35 @@
+## 1.0.5 (2024-07-28)
+
+
+### 🩹 Fixes
+
+- update language server to fix #182
+
+- update vscode-css-languageservice
+
+- support declaring exports without path
+
+## 1.0.4 (2024-07-25)
+
+
+### 🩹 Fixes
+
+- update language server for completion regression fixes
+
+## 1.0.3 (2024-06-16)
+
+
+### 🩹 Fixes
+
+- update language server for completions bugfix
+
+## 1.0.2 (2024-06-15)
+
+
+### 🩹 Fixes
+
+- read setting for maximum color decorators ([25d276b5](https://github.com/wkillerud/some-sass/commit/25d276b5))
+
 ## 1.0.1 (2024-05-01)